Use (a copy of) `tree-model`
This is far simpler than trying to write my own tree although I couldn't get it
to work with mobx due to my inexperience with it.

I've copied in the code for `tree-model` rather than just using it so in the
future I might be able to decorate the internals of the module with mobx
observables and not have to worry about forcing an update to the UI by
arbitrarily updating the `path` state even when it's not required.

Failing this, I may extract this code into a custom hook and ensure I force an
update manually like this:
https://stackoverflow.com/questions/53215285/how-can-i-force-component-to-re-render-with-hooks-in-react