Make the checkbox work (with bugs)
It seems the mobx observable is updating but the UI is not updating correctly
when interacting with the checkbox.

When marking an item as complete, these things should happen:
- The tick should appear (working)
- The progress should animate to 100% (working)
- The item should gain the strike through style (not working)
- The item should be reordered to the bottom (not working)

The other bug is when marking an item as incomplete, the animation does not
occur.

I don't understand mobx well enough to fix this :'(