Make the UI react to changes to the observables
This fixes all the issues outlined in the last commit except the for the
progress not animating back to 0.

I was able to work out this fix by using the following resources:
- https://mobx.js.org/react-integration.html#callback-components-might-require-observer
- https://mobx.js.org/actions.html#asynchronous-actions
- https://mobx.js.org/understanding-reactivity.html

I think the issue was caused by the fact that react-beautiful-dnd uses
function-as-child-components for Draggable and Droppable which result in an
extra layer of function calls that means the observer call does not take affect
on the inner function and as the inner functions are not react components, I
needed to render an `<Observer>` instead of using `observer()`.