Allow items to be deleted
This change involves allowing each node to notify the parent tree object of a
change, which in turn allows the react component that created the tree to
register a callback for when an item is deleted.

This seems like it goes against the way mobx is supposed to work but I don't
understand enough to come up with a better, more "mobx" solution.