Use markdown editor for editing descriptions
The editor works well but rendering the markdown stand-alone does not seem to be
a feature of the library.

The only rendering it does is during preview mode, which in this commit I'm
trying keep preview mode activated while it is in editable mode and while it's
pretty close to doing what I want, codemirror is adding additional height and
the styles are not quite right.

I think a better choice may be to setup a different markdown renderer and use it
for both rendering as per usual and also configure easymde to use the custom
renderer so they still match.
This might make it easier to add fancy features later, like GHF markdown for
table support, checkboxes, etc.