WIP: Attempt a new data structure for firebase
I don't quite know how to sync the existing tree structure of JS objects to
Firebase so this is an attempt at changing how the data is structured.

Transformation steps:
  Firebase: Flattened nodes which children_id arrays
  ProjectStore: Mobx store with array of flat nodes and root of tree structure
  ProjectNode: Observable nodes of a tree structure (like before)
  React: Observes the tree structure

This means everytime Firebase updates, it will recreate the entire tree
structure but hopefully that's not too expensive!

Alternatively, I may be able to use the child events from Firebase to just
update nodes that have changed but not sure this will help because the
references to other nodes will still need to be updated somehow ?