WIP: Efficiently syncing a flat list of observables
A flat list structure of observables may be needed in addition to the tree structure to allow efficient updates to the flat list and the tree to be a computer property based off that flat list.

This commit also experiements with value callbacks vs child callbacks from firebase.
More investigation is required to understand the optimal approach though.