Add child nodes at the end of the sort order
This requires jumping through hoops as Firebase only stores objects and when treating those objects as an array and pushing to it, it always adds it to the end because the auto-generated keys automatically collate in chronological order.

However, when the user decides to reorder tasks in a list, these chronologically ordered keys will not be good enough so I have chosen to use the value as a sort key.

This means:
- Adding new items needs to have a larger sort key so it is added to the bottom
- Rearranging child nodes needs to update all items with an appropriate sort key