Store new items with the next index as a sort key
This assumes the children always have sequential numbers as values, if there are gaps this will liekly double up ids.

This approach is much easier because it requires less effort when there are no children_ids keys, otherwise this logic would need to be gaurded again.

This implementation also does gaurd against it but it's similar.