Make the ProjectStore not based on a single project
The ProjectStore can now be instantiated without the need to have a projectId
which will make it easier to use this for creating new projects.

The ProjectStore can now create Project objects, which are aware of how to get
the details of a project including creating ProjectNode objects which are
accessed in a tree-like fashion.

The RootStore will hopefully make it easy to allow the ProjectStore to fetch a
list of projects belonging to the user too.