Attempt to cache previously logged in state
This caches the user to localStorage so next time it can preemptively render the
correct screen and prevent jank, however I am getting this error due to a
mismatch of what is rendered on the server vs the client:

> Warning: Text content did not match. Server: "It's easy — no login required!"
> Client: "Good to see you again, create a new project here."

This isn't perfect but it's a bit better than it was.
I may need to investigate some sort of loading state which it's still loading.